youtube server 2
Sửa đổi cấu trúc YoutubeSerer2 cx sử dụng đối tượng Option

+++ youtube/youtube_download_server_2.py
@@ -8,31 +8,28 @@ from pytubefix import cipher
 from tqdm import tqdm  # Hiển thị tiến trình tải video
 import sys
 from message.message import Message
-from youtube_key import YoutubeKey
+from youtube_key import YoutubeKey, Option
+
 # "AIzaSyC6XAJGYvbDHiLYFGpv8BpUz9PRSNwIQEA"
 class YoutubeServerTwo:
 
     _process = None
-    def __init__(self, process, API_KEY ):
-        self.VIDEO_TEMP_PATH = './settings/video_temp'
-        self.VIDEO_DOWNLOAD_PATH = './video_download'
-        self.API_KEY = API_KEY
+    def __init__(self, process, option  = Option()):
         self._process = process
-        os.makedirs(self.VIDEO_TEMP_PATH, exist_ok=True)
-        os.makedirs(self.VIDEO_DOWNLOAD_PATH, exist_ok=True)
+        self.option = option
         
 
     #================================ Lấy data video ===========================
 
-    def get_channel_id(self,channel_url):
+    def get_channel_id(self):
         """Trích xuất ID kênh từ URL kênh YouTube"""
-        match = re.search(r"@([\w-]+)", channel_url)
+        match = re.search(r"@([\w-]+)", self.option.channel_url)
         if not match:
             return None
         channel_name = match.group(1)
 
         # Gọi API để lấy ID kênh
-        url = f"https://www.googleapis.com/youtube/v3/search?part=id&q={channel_name}&type=channel&key={self.API_KEY}"
+        url = f"https://www.googleapis.com/youtube/v3/search?part=id&q={channel_name}&type=channel&key={self.option.API_KEY}"
         response = requests.get(url).json()
 
         if "items" in response and response["items"]:
@@ -55,7 +52,7 @@ class YoutubeServerTwo:
 
     def get_video_details(self,video_ids):
         """Lấy thời lượng video để phân loại"""
-        url = f"https://www.googleapis.com/youtube/v3/videos?part=contentDetails&id={','.join(video_ids)}&key={self.API_KEY}"
+        url = f"https://www.googleapis.com/youtube/v3/videos?part=contentDetails&id={','.join(video_ids)}&key={self.option.API_KEY}"
         response = requests.get(url).json()
         
         video_info = {}
@@ -67,18 +64,17 @@ class YoutubeServerTwo:
         
         return video_info
 
-    def get_videos(self,channel_id,option_list_video):
+    def get_videos(self,channel_id):
         """Lấy danh sách video từ kênh theo các tiêu chí"""
         
         order_mapping = { 1: "viewCount", 2: "date", 3: "date"}
-        order = order_mapping.get(int(option_list_video["video_sort"]), "date")
-
+        order = order_mapping.get(int(self.option.video_sort), "date")
         video_list = []
         next_page_token = ""
         
         while True:
             # 🔹 Gọi API lấy video
-            url = f"https://www.googleapis.com/youtube/v3/search?part=snippet&channelId={channel_id}&key={self.API_KEY}&order={order}&maxResults=50&type=video&pageToken={next_page_token}"
+            url = f"https://www.googleapis.com/youtube/v3/search?part=snippet&channelId={channel_id}&key={self.option.API_KEY}&order={order}&maxResults=50&type=video&pageToken={next_page_token}"
             response = requests.get(url).json()
             
             if "items" not in response:
@@ -91,27 +87,27 @@ class YoutubeServerTwo:
                 duration = video_durations.get(video_id, 0)
 
                 # 🔹 Lọc video theo loại mong muốn
-                if  option_list_video["video_duration"]== YoutubeKey.VideoDuaration.VIDEODURATION_20MINUTE and duration <= 1200:
+                if  self.option.video_duration== YoutubeKey.VideoDuaration.VIDEODURATION_20MINUTE and duration <= 1200:
                     continue  # Loại bỏ video ngắn hơn 20 phút
-                if option_list_video["video_duration"] == YoutubeKey.VideoDuaration.VIDEODURATION_FROM4TO20MINUTE and (duration < 240 or duration > 1200):
+                if self.option.video_duration == YoutubeKey.VideoDuaration.VIDEODURATION_FROM4TO20MINUTE and (duration < 240 or duration > 1200):
                     continue  # Loại bỏ video không trong khoảng 4-20 phút
-                if option_list_video["video_duration"] == YoutubeKey.VideoDuaration.VIDEODURATION_FROM1TO4MINUTE and duration >= 240:
+                if self.option.video_duration == YoutubeKey.VideoDuaration.VIDEODURATION_FROM1TO4MINUTE and duration >= 240:
                     continue  # Loại bỏ video dài hơn 4 phút
-                if option_list_video["video_duration"] == YoutubeKey.VideoDuaration.VIDEODURATION_UNDER1MINUTE and duration >= 60:
+                if self.option.video_duration == YoutubeKey.VideoDuaration.VIDEODURATION_UNDER1MINUTE and duration >= 60:
                     continue  # Loại bỏ video không phải Shorts
 
                 video_list.append(video_id)
 
             # 🔹 Nếu chỉ lấy số lượng nhất định, dừng khi đủ
-            if int(option_list_video["count"]) > 0 and len(video_list) >= int(option_list_video["count"]):
+            if int(self.option.count) > 0 and len(video_list) >= int(self.option.count):
                 break
 
             # 🔹 Nếu có trang tiếp theo, tiếp tục lấy
             next_page_token = response.get("nextPageToken", "")
-            if not next_page_token or int(option_list_video["count"]) == -1:
+            if not next_page_token or int(self.option.count) == -1:
                 break
 
-        return video_list[0:int(option_list_video["count"])] if int(option_list_video["count"]) == -1 else video_list
+        return video_list[0:int(self.option.count)] if int(self.option.count) == -1 else video_list
 
 
     #================================ Xử lý Video ===========================
@@ -153,14 +149,17 @@ class YoutubeServerTwo:
         """Xóa ký tự đặc biệt trong tên file."""
         return re.sub(r'[\\/*?:"<>|]', "", filename)
 
-    def download_video_and_audio(self,video_id, resolution='1080p'):
+    def download_video_and_audio(self,video_id):
         """Tải video từ YouTube và ghép âm thanh."""
         url = f'https://youtube.com/watch?v={video_id}'
         # url = f'https://www.youtube.com/shorts/{video_id}'
+
+        resolution = "1080p" if self.option.video_quatity == YoutubeKey.VideoQuality.VideoFormat[0] or self.option.video_quatity == YoutubeKey.VideoQuality.VideoFormat[2] else "360p"
+
         try:
             yt = YouTube(url)
             safe_title = self.sanitize_filename(yt.title)
-            output_file = os.path.join(self.VIDEO_DOWNLOAD_PATH, f"{safe_title}.mp4")
+            output_file = os.path.join(self.option.output_path, f"{safe_title}.mp4")
             
             if os.path.exists(output_file):
                 print(f"Bỏ qua {safe_title} vì đã tải xuống.")
@@ -168,25 +167,31 @@ class YoutubeServerTwo:
             
             video_stream = yt.streams.filter(file_extension='mp4', res=resolution, only_video=True).first()
             audio_stream = yt.streams.filter(only_audio=True).first()
-
+            
             if video_stream and audio_stream:
-                video_file = video_stream.download(output_path=self.VIDEO_TEMP_PATH, filename='video.mp4')
-                audio_file = audio_stream.download(output_path=self.VIDEO_TEMP_PATH, filename='audio.mp4')
-                
-                command = f'ffmpeg -hide_banner -loglevel error -i "{video_file}" -i "{audio_file}" -c:v copy -c:a aac "{output_file}"'
-                subprocess.run(command, shell=True, check=True)
-                
-                os.remove(video_file)
-                os.remove(audio_file)
-                print(f"✅ Tải xong: {safe_title}")
+                video_file = ""
+                audio_file = ""
+                if self.option.video_quatity == YoutubeKey.VideoQuality.VideoFormat[0]:
+                    video_file = video_stream.download(output_path=self.VIDEO_TEMP_PATH, filename='video.mp4')
+                    audio_file = audio_stream.download(output_path=self.VIDEO_TEMP_PATH, filename='audio.mp4')
+                    command = f'ffmpeg -hide_banner -loglevel error -i "{video_file}" -i "{audio_file}" -c:v copy -c:a aac "{output_file}"'
+                    subprocess.run(command, shell=True, check=True)
+                    
+                    os.remove(video_file)
+                    os.remove(audio_file)
+                    print(f"✅ Tải xong: {safe_title}")
+                if self.option.video_quatity == YoutubeKey.VideoQuality.VideoFormat[2] or self.option.video_quatity == YoutubeKey.VideoQuality.VideoFormat[4]:
+                    video_file = video_stream.download(output_path=self.VIDEO_TEMP_PATH, filename='video.mp4')
+                if self.option.video_quatity == YoutubeKey.VideoQuality.VideoFormat[3] or self.option.video_quatity == YoutubeKey.VideoQuality.VideoFormat[5]:
+                    audio_file = audio_stream.download(output_path=self.VIDEO_TEMP_PATH, filename='audio.mp4')
             else:
                 print(f"⚠️ Không tìm thấy stream phù hợp cho {safe_title}")
         except Exception as e:
             print(f'❌ Lỗi tải video {video_id}: {e}')
 
-    def run(self, option_list_video):
+    def run(self):
 
-        channel_id = self.get_channel_id(option_list_video["url"])
+        channel_id = self.get_channel_id()
 
         if not channel_id:
             if self._process:
@@ -195,18 +200,7 @@ class YoutubeServerTwo:
             if self._process:
                 self._process(Message.NotificationType.NOTIFICATION, f"✅ ID kênh: {channel_id}")
 
-            video_type = option_list_video["video_duration"]
-            
-            sort_by = option_list_video["video_sort"]
-            type_craw = int(option_list_video["count"])
-
-            count_video = 50  # Giá trị mặc định
-            if type_craw < 0:
-                count_video = 50
-
-            # 🔹 Lấy danh sách video
-            # video_type, sort_by, type_craw, count_video
-            video_list = self.get_videos(channel_id, option)
+            video_list = self.get_videos(channel_id)
 
             print(f"📌 Danh sách video: {video_list}")
